GUI Changes
Settings.txt file was created to use custom syntax to add and modify the fields available to the user. the file has a template currently inside that does throw an error, but shows how to create fields. The program ID and it's descriptor text line are currently related through a map. The program ID and the string arraylist of all text lines that relate to that ID are stored in a map. The Program ID and the actual field input components are related by a map. The GUI is running insdie of a scroll Pane with a viewport on a JPanel which is following a YAxis Box Layout. Each Field is added to a split pain, with the Nick on the left and the Input on the right.