Adding/removing fields is pin protected
In order to make it possible to have this in a shared drive for all teachers but keep it secure so they don't accidentally edit fields they shoulldn't, we have added a pin prompt. It's not that secure, and can be cracked just by looking at the code, but it's just intended to make sure teachers don't change things on accident.